A few minor updates and changes to Zatch
I have some new code which is not done, I'm only leaving it in there for
the purpose of allowing people to try and get it to work. I've also
updated the Auto-op list I'm planning to add in a few more features as
well when I get around to it.