File fix! Also changed the way .rrtoggle works
I accidentally pushed the wrong files, so here. Here's more info on the
change of .rrtoggle I changed the way .rrtoggle works from an int(using 0
and 1) to a boolean, this way it's more efficient in the long run, besides
I don't need numbers to toggle the state, I just need to turn it on and
off.