First Rewite
I have began rewriting Zatch. The Rewrite should remove redudant code and
organize the code better. Also certain Hardcoded things will be changeable
in a config file, right now it's possible to connect to 1 server and 1
channel through the config file(which you need to manually make), with the
sever, then a new line and the Channel on that new line. I'm working on
making that portion of the config file explained, with autogeneration, and
limitless channels to join on autoconnect(although I can't do that with
Servers due to the way PIRC handels servers(1 server per bot instance).