Config File setup and working
Alright so the config file is completely setup for the time being. There
may be some difference in the loading of configs and the way the configs
are saved, but for now It's fine. Right now all files are loaded into
Zatch properly from the looks of it, however I'm still testing. I have yet
to impliment the auto-op feature again.