Fixed a small error
I fixed an error in the configs for formatting and the loading of the
config. The error would cause it to incorrectly detect the master, I also
changed the layout of the config to save on lines of Code. Also the
auto-op feature is being set-up with the code for the ops ready. Right now
you require hostnames as well in order to prevent people from stealing the
names of someone.