Features are finished and implimented
I have added the config file completely, as well as rewriting and
organizing the code. Among new features are: The ability to change the
location of the logs as well as toggling logging on or off without going
through the code, the ability to change the name of Zatch through configs,
the ability to set Zatch to auto-op people in the config(as well as check
their hostnames if you activate it), and to toggle the authentication for
Zatch, the ability to change the "master" so he can respond to your
commands without editing the code, autoconfig generation, the ability to
change the server throught the config.

I also removed the Random Responses for right now, I may impliment them
later but I probably won't. Also this is the final commit until I merge
them, unless a major bug is found(I'm going to be keeping this in this
branch for a few days and moving Zatch to a dedicated area to test him on
for a bit.