Fixed a small error
I fixed a minor error which would prevent all of the trigger statements
and the xchan command from working(unless you were the defined Master in
the config). This was due to a forgotten Brace(AKA Curly Bracket), which I
then placed further down in the code on accident. This is now fixed