Fixed Grammer for master Help command
I have edited the Help command for master to be gramatically correct. Also
there is a peice of code for a Operator Add command, but it's incomplete I
will remove it before the merge into Master and then branch master into a
new branch for working on this feature, at the moment though I'm leaving
it in there(I have commented it out).