Changed some comments, added license
Alright so I changed some comments and stuff. Since this is using Pircbot
I kinda have to license the bot under GNU GPL v2.0 as such, I have now
done so. Also this should be the final version of Zatch, unless there is a
bug or something. I might be working on another bot, but not atm.